+++ src/index.html
@@ -23,6 +23,9 @@
         <th>Century</th>
       </tr>
     </table>
-    <script src="./scripts/main.js"></script>
+    <script
+      type="module"
+      src="./scripts/main.js"
+    ></script>
   </body>
 </html>
